requirements.txt
fastapi~=0.116.1
pydantic~=2.11.9
starlette~=0.47.3
sqlmodel~=0.0.24
sqlalchemy~=2.0.43
pydantic_core
alembic~=1.16.5
langchain_openai
langchain_community
langchain~=0.3.27
dashscope
unstructured~=0.18.15
pytest~=8.4.2
pandas~=2.3.2
datasets~=4.1.1
ragas~=0.3.5
passlib~=1.7.4
PyJWT~=2.10.1
celery~=5.5.3
redis~=5.2.1